Fix types

+++ programs/agora-court/src/lib.rs
@@ -17,7 +17,7 @@ pub mod agora_court {
     pub fn create_dispute(
         ctx: Context<CreateDispute>,
         users: Vec<Pubkey>,
-        order_price: u32,
+        order_price: u64,
         config: DisputeConfiguration,
     ) -> Result<()> {
         instructions::create_dispute(ctx, users, order_price, config)
